ref(api): Sentry error submodule
